producer_app/requirements.txt
websocket-client
confluent-kafka